Dashboard Views and Controllers
*Make sure that you’re .env file in the project root has the correct APP_URL set. The way I set the nav links up requires this to be set and correct

Dashboard Styles – Continued building the dashboard, added the paws logo and search bar

Moved all models to a model folder in the app directory for better organization. You can access them in a class via use App\Models\model-name

Created a new view and controller method to view all dogs in the database. All records are displayed out in a table.

Created a new dog view and wrote an ajax function to add the new dog to the database. I also added in a pop up success modal upon saving that allows the user to add another new dog or return to the list of dogs.

Created an edit dog view but we still need to finish building the edit view and required controllers.

Updated the dogs table to store the breed as a string instead of the bread id to simplify the querying process.